+++ aegisub/audio_provider_ram.h
@@ -55,5 +55,5 @@ public:
 	RAMAudioProvider(AudioProvider *source);
 	~RAMAudioProvider();
 
-	void GetAudio(void *buf, long long start, long long count);
+	void GetAudio(void *buf, int64_t start, int64_t count);
 };
